Move integration tests to a separate directory
